config-provided "up_service" feature
