+++ irods_docker_files/docker_compose_CI_with_client.py
@@ -10,22 +10,15 @@ import sys
 import copy
 import json
 import time
-from pprint import pformat
+
 from os.path import join
-import docker
 import compose.cli.command
 
-class testgen:
-    def __init__(self,file_,tm=0.15):
-        self.file = (file_ if getattr(file_,'read',None) else open(file_,"rb"))
-        self.tm = tm
-    def __iter__(self):
-        with self.file as f:
-          x = '.'
-          while x:
-            x = f.read(2)
-            time.sleep(self.tm)
-            if x: yield x
+from docker_compose_ci_util import (testgen, readgen,
+                                    run_build_containers,
+                                    get_ordered_dependencies)
+
+# -------------------------------- Test code
 
 def test():
     import io
@@ -38,125 +31,6 @@ def test():
         ident_ = 'mhyident'
     )
 
-
-def compile_EOL_regex(delimiter): return re.compile( b'('       # parentheses mean that we retain the delimiters
-                                                    + delimiter
-                                                    + b')' )
-
-class readgen:
-
-    MAX_LINE_ELEMENTS = 256
-    EOL_splitters = { delim:compile_EOL_regex(delim) for delim in [b'\n',b'\r\n'] }
-    EOL_splitters[b''] = re.compile(b'(.+)',re.DOTALL)
-    class IllegalDelimiter(RuntimeError): pass
-
-    class dummyLock:
-        def __enter__(self): pass
-        def __exit__(self,*x): pass
-
-    def __init__(self, files_ , file_locks_ = ()):
-        self.buf = []
-        self.files = files_
-        self.fileLocks = dict(file_locks_)
-        self.ident = ''
-
-    # Get a piece of the container log_stream's output to be printed to console.
-    # (Usually will consist of zero or more lines)
-
-    def get_buffered_content(self, new_chars=b'', delim = b'\n'): # b'\n' => output only whole lines
-                                                                  # b''   => output all of buffer (flush)
-        log_lines = b''
-        if not isinstance(new_chars,bytes):
-            new_chars = str(new_chars).encode('utf-8')
-        self.buf.append(new_chars)
-        if delim in new_chars or len(self.buf) > self.MAX_LINE_ELEMENTS:
-            regexSplitter = self.EOL_splitters.get(delim)
-            if regexSplitter is None:
-                if isinstance(delim,bytes): regexSplitter = self.compile_EOL_regex(delim)
-                else:                       raise self.IllegalDelimiter('cannot make delimiter of: {!r}'.format(delim))
-            buf = regexSplitter.split(b''.join(self.buf)) + [b'']
-            cut_offs = 0; last = 1                            # Calculate the negative end-offset of the
-            if isinstance(delim,bytes) and len(delim) >= 1:   # last chunk containing a 'delim' character.
-                for chunk in reversed( buf[-3:-1] ):          # This will always be 1 for a flush operation.
-                    last += 1
-                    if chunk.startswith(delim):
-                        cut_offs=1
-                        break
-            log_lines += b''.join(buf[:-last])
-            self.buf = list(filter(None,buf[-last+cut_offs:]))
-        else:
-            pass
-#######
-# dwm #
-      # if 'runner' in  self.ident:
-      #     log_lines = '*** dwm *** LEN of els in line buffer = {!r}'.format([len(x) for x in self.buf]).encode('utf-8')
-      # # dwm DEBUG -- # print('\t self.buf=', self.buf,file=sys.stderr)
-#######
-        return log_lines
-
-    # Thread entry point.
-
-    def __call__(self, log_generator, ident_ = None) :
-        def stream_out(Lines):
-            if not Lines: return
-            for f in self.files:
-                with self.fileLocks.get(f,self.dummyLock()):
-                    for line in Lines:
-                        f.write("(" + self.ident + ") -- | " + line.decode("utf-8") + "\n")
-        if ident_ is None: ident_ = '<{}>'.format(id(self))
-        self.ident = ident_
-        for chunk in log_generator:
-            stream_out (self.get_buffered_content(chunk, delim = b'\n').splitlines())
-        stream_out (self.get_buffered_content(b'', delim = b'').splitlines())
-
-
-def run_build_containers( project, service_names_in_order ):
-    service_lookup = { s.name:s for s in project.services }   
-    # iterate through the named services, running the build process and CMD phase for each
-    # Output should appear on the console for all parts built or commands run.
-    for name in service_names_in_order:
-        project.build(service_names = [name])
-        s = service_lookup[name]
-        c = s.create_container()
-        dclient = docker.client.from_env()
-        for network_name in s.networks:
-            try:
-                dclient.networks.get(network_name)
-            except docker.errors.NotFound:
-                dclient.networks.create(network_name)
-        if not getattr(c,'log_stream',None):
-            c.attach_log_stream()
-        c.start()
-        for i in c.log_stream:
-            print ('>> '+i.decode('utf8'))
-        w = c.wait()
-        print(s.name,'returned',w)
-
-
-def get_ordered_dependencies(svcn,prj,ser=None,pdict=None,order=(),seen=()):
-    '''Given:
-       - svcn, the name of the service for the main CI test container (whose exit status determines success).
-       - prj, the compose project returned by get_project().
-
-    Calculate list of names of all svcn's dependents that must be built (inclusive of svcn, which will be last).
-    '''
-    if pdict is None:
-        pdict = {svc.name:svc for svc in prj.services}
-    if order == ():
-        tmp_order = dict(order)  # initialize ordering dictionary at root of recursion
-    else:
-        tmp_order = order
-    seen = dict(seen) if seen == () else seen
-    ser = [0] if ser is None else ser
-    dep_names = pdict[svcn].get_dependency_names()
-    tmp_order[svcn] = ser[0]
-    ser[0] += 1
-    seen[svcn]=1
-    for name in dep_names:
-        if not seen.get(name):
-            get_ordered_dependencies(name,prj,ser,pdict,tmp_order)
-    if order == ():
-        return [k for k,v in sorted(tmp_order.items(), key=lambda _:_[1],reverse=True)]
 
 
 def main() :
@@ -198,8 +72,10 @@ def main() :
     import importlib
     sys.path.insert(0, Args.local_repository)
 
-    # Modify the python3 command invoking this script
-    # To insert new override KEY=VALUE pairs for modifiers_for_config, for example
+    # You can modify the python3 command invoking this script from the Jenkins config.xml pipeline section, in
+    # order to insert new override KEY=VALUE pairs for modifiers_for_config.
+    # For example:
+    #
     #    def build_cmd = 'env irods_package_dir="'+IRODS_PACKAGE_DIR+'" ' +  // IRODS_PACKAGE_DIR entry box value
     #                             '    X="' + 'HELLO' + '" ' +               // literal value "HELLO"
     #                             '    DOTENV_INJECT_KEYS=irods_package_dir,X ' +
@@ -252,7 +128,7 @@ class CI_client_interface (object):
 
 
     def run_and_wait_on_client_exit( self,
-                                     name_pattern = '[-_]client-runner[-_]',
+                                     name_pattern = 'client[-_]runner',
                                      rgx_flags    = re.IGNORECASE,
                                      import_vars  = True ):
 
@@ -269,29 +145,43 @@ class CI_client_interface (object):
         if  proj is None:
             proj = compose.cli.command.get_project( self.compose_path )
 
-        build_order = self.config.get("build_services_in_order",[])
+        build_order = self.config.get("build_services_in_order")  # should be 'None' if no build/prepare services
 
+        Entire_Project_Built = False
         if build_order is None:
             proj.build()                 # build all services from docker-compose.yml
             Entire_Project_Built = True
         else:
-                                         # build and run named services (to prepare/build for client run)
-            run_build_containers (proj, [build_order] if isinstance (buildorder,str) else list(build_order))
-            Entire_Project_Built = False
+            # build and run named services (to prepare/build for client run)
+            # if empty (specified as '' or []) build nothing in the project
+            if build_order:
+                run_build_containers(proj,
+                                     [build_order] if isinstance (build_order,str) else list(build_order))
 
         #  -- TODO  --
         #  Determine which container to wait on for status code.
         #  Usually "client-runner" or similar
-        # 
-        #  If entire project wasn't previously built, iterate through names of dependent services
-        #  and build them. then initiate the named client container
+        #  Use "depends_on:" in the docker-compose.yml to specify run dependencies of the main container and thus which
+        #    services run in the up() phase.
 
-        containers = self.compose_prj = proj.up()
+        main_service = self.config.get("up_service") # is the None object, if name not provided
 
-        # -- Match client container by name for status reporting
+        if not Entire_Project_Built:
+            if not main_service:
+                print ("ERROR - A comprehensive docker-compose project build was not performed because build stages were given in the config.")
+                print ("        But an 'up_service' name was not given, which could result in running out-of-date services.  Exiting.... ")
+                exit(1)
+            proj.build( service_names=get_ordered_dependencies(proj,main_service) )
+
+        containers = self.compose_prj = proj.up(service_names = [main_service] if main_service else None)
+
+        # -- Match client container by a default name pattern if not given, else choose by config-provided name
         #
-        rgx = re.compile(name_pattern, rgx_flags)
-        status_containers = [ c for c in containers if rgx.search(c.name) ]
+        if main_service:
+            status_containers = [ c for c in containers if c.name_without_project.startswith(main_service) ]
+        else:
+            rgx = re.compile(name_pattern, rgx_flags)
+            status_containers = [ c for c in containers if rgx.match(c.name_without_project) ]
 
         if len(status_containers) != 1:
             raise RuntimeError("Exactly one container can be waited on for CI status, but the"
@@ -346,5 +236,6 @@ class CI_client_interface (object):
 if __name__ == '__main__':
     if sys.argv[1:] == ['-test']:
         test()
+        p = compose.cli.command.get_project('.')
     else:
         main()
